Moving the window implementation into its own package.
