+++ src/main/java/net/eugenpaul/jlexi/appl/action/BoldActivate.java
@@ -1,4 +1,4 @@
-package net.eugenpaul.jlexi.window;
+package net.eugenpaul.jlexi.appl.action;
 
 import java.beans.PropertyChangeEvent;
 
@@ -11,7 +11,7 @@ import net.eugenpaul.jlexi.design.listener.MouseEventAdapter;
 import net.eugenpaul.jlexi.utils.event.MouseButton;
 
 @AllArgsConstructor
-public class ItalicActivate implements MouseEventAdapter {
+public class BoldActivate implements MouseEventAdapter {
 
     private final String name;
     private final Button button;
@@ -23,9 +23,9 @@ public class ItalicActivate implements MouseEventAdapter {
             this.button.setState(ButtonState.PUSHED);
             this.button.redraw();
 
-            controller.propertyChange(new PropertyChangeEvent(//
-                    name, //
-                    ViewPropertyChangeType.CURSOR_SET_FORMAT_ITALIC.getTypeName(), //
+            this.controller.propertyChange(new PropertyChangeEvent(//
+            this.name, //
+                    ViewPropertyChangeType.CURSOR_SET_FORMAT_BOLD.getTypeName(), //
                     null, //
                     Boolean.TRUE //
             ));
@@ -33,9 +33,9 @@ public class ItalicActivate implements MouseEventAdapter {
             this.button.setState(ButtonState.NORMAL);
             this.button.redraw();
 
-            controller.propertyChange(new PropertyChangeEvent(//
-                    name, //
-                    ViewPropertyChangeType.CURSOR_SET_FORMAT_ITALIC.getTypeName(), //
+            this.controller.propertyChange(new PropertyChangeEvent(//
+            this.name, //
+                    ViewPropertyChangeType.CURSOR_SET_FORMAT_BOLD.getTypeName(), //
                     null, //
                     Boolean.FALSE //
             ));
