+++ src/main/java/net/eugenpaul/jlexi/appl/action/PasteAction.java
@@ -1,0 +1,5 @@
+package net.eugenpaul.jlexi.appl.action;
+
+public class PasteAction {
+    
+}
