+++ src/main/java/net/eugenpaul/jlexi/appl/action/RedoAction.java
@@ -1,0 +1,5 @@
+package net.eugenpaul.jlexi.appl.action;
+
+public class RedoAction {
+
+}
