+++ src/main/java/net/eugenpaul/jlexi/appl/action/UndoAction.java
@@ -1,0 +1,5 @@
+package net.eugenpaul.jlexi.appl.action;
+
+public class UndoAction {
+    
+}
