+++ src/main/java/net/eugenpaul/jlexi/appl/action/UndoActivate.java
@@ -1,4 +1,4 @@
-package net.eugenpaul.jlexi.window;
+package net.eugenpaul.jlexi.appl.action;
 
 import lombok.AllArgsConstructor;
 import net.eugenpaul.jlexi.controller.ModelController;
@@ -6,13 +6,13 @@ import net.eugenpaul.jlexi.design.listener.MouseEventAdapter;
 import net.eugenpaul.jlexi.utils.event.MouseButton;
 
 @AllArgsConstructor
-public class PasteActivate implements MouseEventAdapter {
+public class UndoActivate implements MouseEventAdapter {
 
     private final String name;
     private final ModelController controller;
 
     @Override
     public void mouseClicked(Integer mouseX, Integer mouseY, MouseButton button) {
-        this.controller.paste(name);
+        this.controller.undo(name);
     }
 }
