+++ src/main/java/net/eugenpaul/jlexi/appl/impl/swing/frame/ImagePanel.java
@@ -1,4 +1,4 @@
-package net.eugenpaul.jlexi.window.impl.swing.frame;
+package net.eugenpaul.jlexi.appl.impl.swing.frame;
 
 import java.awt.Graphics;
 import java.awt.Image;
