+++ src/main/java/net/eugenpaul/jlexi/appl/impl/swing/frame/KeyListener.java
@@ -1,4 +1,4 @@
-package net.eugenpaul.jlexi.window.impl.swing.frame;
+package net.eugenpaul.jlexi.appl.impl.swing.frame;
 
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
