+++ src/main/java/net/eugenpaul/jlexi/appl/impl/swing/frame/MouseListner.java
@@ -1,4 +1,4 @@
-package net.eugenpaul.jlexi.window.impl.swing.frame;
+package net.eugenpaul.jlexi.appl.impl.swing.frame;
 
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
