--- src/main/java/net/eugenpaul/jlexi/component/text/action/PasteAction.java
@@ -1,5 +1,0 @@
-package net.eugenpaul.jlexi.component.text.action;
-
-public class PasteAction {
-    
-}
